塊の当たり判定の変更、ファイル移動、BGMとSEの追加、命名規則の変更、コメントアウト
塊のcharacterControllerを削除し、rigidBodyに変更しました。
