+++ GameTemplate/Game/Actor/Object/StageObjectBase.cpp
@@ -1,6 +1,6 @@
 #include "stdafx.h"
 #include "StageObjectBase.h"
-#include "Actor/Sphere.h"
+#include "Actor/Sphere/Sphere.h"
 
 
 namespace
