+++ GameTemplate/Game/Actor/Sphere/SphereScaleManager.cpp
@@ -1,0 +1,1 @@
+#include "stdafx.h"
