+++ GameTemplate/Game/Actor/Sphere/SphereScaleManager.h
@@ -1,0 +1,20 @@
+/**
+ * SphereScaleManager.h
+ * 塊の大きさを管理するクラス
+ */
+
+#pragma once
+#include "Sphere.h"
+
+
+class SphereScaleManager : public Sphere 
+{
+	/**
+	 * @brief コンストラクタ
+	 */
+	SphereScaleManager() {};
+	/**
+	 * @brief デストラクタ
+	 */
+	~SphereScaleManager() {};
+};
