--- GameTemplate/Game/Actor/SphereScaleManager.cpp
@@ -1,1 +1,0 @@
-#include "stdafx.h"
