--- GameTemplate/Game/Actor/SphereScaleManager.h
@@ -1,20 +1,0 @@
-/**
- * SphereScaleManager.h
- * 塊の大きさを管理するクラス
- */
-
-#pragma once
-#include "Sphere.h"
-
-
-class SphereScaleManager : public Sphere 
-{
-	/**
-	 * @brief コンストラクタ
-	 */
-	SphereScaleManager() {};
-	/**
-	 * @brief デストラクタ
-	 */
-	~SphereScaleManager() {};
-};
