+++ GameTemplate/Game/Game.vcxproj
@@ -228,9 +228,9 @@
     <ClCompile Include="Actor\Object\MovableObject.cpp" />
     <ClCompile Include="Actor\Object\StageObjectBase.cpp" />
     <ClCompile Include="Actor\Object\StaticObject.cpp" />
-    <ClCompile Include="Actor\Player.cpp" />
-    <ClCompile Include="Actor\Sphere.cpp" />
-    <ClCompile Include="Actor\SphereScaleManager.cpp" />
+    <ClCompile Include="Actor\Sphere\Player.cpp" />
+    <ClCompile Include="Actor\Sphere\Sphere.cpp" />
+    <ClCompile Include="Actor\Sphere\SphereScaleManager.cpp" />
     <ClCompile Include="Collision\CollisionManager.cpp" />
     <ClCompile Include="Game.cpp" />
     <ClCompile Include="GameManager.cpp" />
@@ -258,9 +258,9 @@
     <ClInclude Include="Actor\Object\MovableObject.h" />
     <ClInclude Include="Actor\Object\StageObjectBase.h" />
     <ClInclude Include="Actor\Object\StaticObject.h" />
-    <ClInclude Include="Actor\Player.h" />
-    <ClInclude Include="Actor\Sphere.h" />
-    <ClInclude Include="Actor\SphereScaleManager.h" />
+    <ClInclude Include="Actor\Sphere\Player.h" />
+    <ClInclude Include="Actor\Sphere\Sphere.h" />
+    <ClInclude Include="Actor\Sphere\SphereScaleManager.h" />
     <ClInclude Include="Collision\CollisionManager.h" />
     <ClInclude Include="Game.h" />
     <ClInclude Include="GameManager.h" />
