+++ GameTemplate/k2EngineLow/k2EngineLow.vcxproj
@@ -317,6 +317,7 @@
     <ClCompile Include="physics\BoxCollider.cpp" />
     <ClCompile Include="physics\CapsuleCollider.cpp" />
     <ClCompile Include="physics\CharacterController.cpp" />
+    <ClCompile Include="physics\CompoundCollider.cpp" />
     <ClCompile Include="physics\DebugWireframe.cpp" />
     <ClCompile Include="physics\PhysicsGhostObject.cpp" />
     <ClCompile Include="physics\MeshCollider.cpp" />
@@ -409,6 +410,7 @@
     <ClInclude Include="physics\CapsuleCollider.h" />
     <ClInclude Include="physics\CharacterController.h" />
     <ClInclude Include="physics\CollisionAttr.h" />
+    <ClInclude Include="physics\CompoundCollider.h" />
     <ClInclude Include="physics\DebugWireframe.h" />
     <ClInclude Include="physics\PhysicsGhostObject.h" />
     <ClInclude Include="physics\ICollider.h" />
