+++ GameTemplate/k2EngineLow/physics/CompoundCollider.cpp
@@ -1,0 +1,8 @@
+/*!
+ * @brief	カプセルコライダー。
+ */
+
+#include "k2EngineLowPreCompile.h"
+#include "Physics.h"
+#include "CompoundCollider.h"
+
